refactor overpass code from _downloader module to new _downloader_overpass module
